pr6.3

+++ delete-plan.php
@@ -1,5 +1,9 @@
 <?php
 require_once 'secure.php';
+if (!Helper::can('admin') && !Helper::can('manager')) {
+    header('Location: 404.php');
+    exit();
+    }
 $id = Helper::clearInt($_GET['id']);
 $idPlan = Helper::clearInt($_GET['idplan']);
 if ((new ScheduleMap())->existsScheduleByLessonPlanId($id) || !(new LessonPlanMap())->delete($id)) {
